Still trying to make translation work.
Everything is translated but the game over screen. Thats because I want to do a typewritter effect on it but I cant figure out how to make that work, but maybe i got it to work on a separated test project. When I get it to work, ill import to this project the new typewritter system.